refactor(pack.move): extract controller and template of component
